Stage_3_and_4/Backend/site1/requirements.txt
Django>=4.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
mysqlclient
pymysql
cryptography
Pillow
requests
gunicorn
